feat(#4): Pull request template
* first version pull_request template

* Improve pull_request template

* Format clarification

Add note that we don't put space between commit type and bracket

We are keeping brackets because they are nice but we want to
be very careful about symbol space in commit messages. However
brackets without space add only 1 symbol comparing with bare
space so it is a reasonable choice!